+++ jira/utils.go
@@ -7,8 +7,9 @@ import (
 	"time"
 
 	"github.com/andygrunwald/go-jira"
-	"github.com/turbot/steampipe-plugin-sdk/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
+	"github.com/turbot/steampipe-plugin-sdk/v3/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin/transform"
 )
 
 func connect(_ context.Context, d *plugin.QueryData) (*jira.Client, error) {
@@ -78,4 +79,56 @@ func convertJiraDate(_ context.Context, d *transform.TransformData) (interface{}
 		return nil, nil
 	}
 	return time.Time(d.Value.(jira.Date)), nil
+}
+
+func buildJQLQueryFromQuals(equalQuals plugin.KeyColumnQualMap, tableColumns []*plugin.Column) string {
+	filters := []string{}
+
+	for _, filterQualItem := range tableColumns {
+		filterQual := equalQuals[filterQualItem.Name]
+		if filterQual == nil {
+			continue
+		}
+
+		// Check only if filter qual map matches with optional column name
+		if filterQual.Name == filterQualItem.Name {
+			if filterQual.Quals == nil {
+				continue
+			}
+
+			for _, qual := range filterQual.Quals {
+				if qual.Value != nil {
+					value := qual.Value
+					switch filterQualItem.Type {
+					case proto.ColumnType_STRING:
+						switch qual.Operator {
+						case "=":
+							filters = append(filters, fmt.Sprintf("\"%s\" = \"%s\"", getIssueJQLKey(filterQualItem.Name), value.GetStringValue()))
+						case "<>":
+							filters = append(filters, fmt.Sprintf("%s != \"%s\"", getIssueJQLKey(filterQualItem.Name), value.GetStringValue()))
+						}
+					case proto.ColumnType_TIMESTAMP:
+						switch qual.Operator {
+						case "=", ">=", ">", "<=", "<":
+							filters = append(filters, fmt.Sprintf("\"%s\" %s \"%s\"", getIssueJQLKey(filterQualItem.Name), qual.Operator, value.GetTimestampValue().AsTime().Format("2006-01-02 15:04")))
+						case "<>":
+							filters = append(filters, fmt.Sprintf("\"%s\" != \"%s\"", getIssueJQLKey(filterQualItem.Name), value.GetTimestampValue().AsTime().Format("2006-01-02 15:04")))
+						}
+
+					}
+				}
+			}
+
+		}
+	}
+
+	if len(filters) > 0 {
+		return strings.Join(filters, " AND ")
+	}
+
+	return ""
+}
+
+func getIssueJQLKey(columnName string) string {
+	return strings.ToLower(strings.Split(columnName, "_")[0])
 }