+++ main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/turbot/steampipe-plugin-jira/jira"
 
-	"github.com/turbot/steampipe-plugin-sdk/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
 )
 
 func main() {